fix subtitle generation (#6)

+++ Models/MatchedEpisode.cs
@@ -1,3 +1,3 @@
 ﻿namespace TvRename.Models;
 
-public record MatchedEpisode(int SeasonNumber, int EpisodeNumber, int Confidence);
+public record MatchedEpisode(int SeasonNumber, int EpisodeNumber, double Confidence);
